PaymentService/Application/PaymentProcessorFactory.cs
using Application.Booking.Dtos;
using Application.MercadoPago;
using Application.Payment.Ports;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Payments.Application
{
    public class PaymentProcessorFactory : IPaymentProcessorFactory
    {
        public IPaymentProcessor GetPaymentProcessor(SupportedPaymentProviders selectedPaymentProvider)
        {
            switch (selectedPaymentProvider)
            {
                case SupportedPaymentProviders.MercadoPago:
                    return new MercadoPagoAdapter();
                default: return new NotImplementedPaymentProvider();
            }
        }
    }
}
